+++ engine/calculators/source/unit_tests/ExoticWeaponCalculator_test.cpp
@@ -19,8 +19,8 @@ TEST(SW_Engine_Calculator_ExoticWeaponCalculator, get_base_damage)
 	ExoticWeaponCalculator ewc;
 	Damage dam = ewc.get_base_damage_for_improvised_exotic_weapon();
 
-	EXPECT_EQ(1, dam.get_num_dice());
-	EXPECT_EQ(2, dam.get_dice_sides());
+	EXPECT_EQ(static_cast<uint>(1), dam.get_num_dice());
+	EXPECT_EQ(static_cast<uint>(2), dam.get_dice_sides());
 	EXPECT_EQ(0, dam.get_modifier());
 }
