+++ engine/maps/source/unit_tests/MapExitUtils_test.cpp
@@ -22,7 +22,7 @@ TEST_F(SW_Engine_Maps_MapExitUtilsFixture, map_tile_exit_added_new_dir)
 
   add_map_tile_exit(tile_exits, direction, tile_coord);
 
-  EXPECT_EQ(1, tile_exits.size());
+  EXPECT_EQ(static_cast<uint>(1), tile_exits.size());
 
   auto m_it = tile_exits.find(direction);
 
@@ -32,7 +32,7 @@ TEST_F(SW_Engine_Maps_MapExitUtilsFixture, map_tile_exit_added_new_dir)
   {
     vector<Coordinate> c = m_it->second;
 
-    EXPECT_EQ(1, c.size());
+    EXPECT_EQ(static_cast<uint>(1), c.size());
 
     EXPECT_EQ(tile_coord, c.at(0));
   }
@@ -51,7 +51,7 @@ TEST_F(SW_Engine_Maps_MapExitUtilsFixture, map_tile_exit_added_to_existing)
 
   auto m_it = tile_exits.find(direction);
 
-  EXPECT_EQ(2, m_it->second.size());
+  EXPECT_EQ(static_cast<uint>(2), m_it->second.size());
 
   // Ensure a particular set of coordinates exists only once for a
   // given direction.
@@ -59,5 +59,5 @@ TEST_F(SW_Engine_Maps_MapExitUtilsFixture, map_tile_exit_added_to_existing)
 
   m_it = tile_exits.find(direction);
 
-  EXPECT_EQ(2, m_it->second.size());
+  EXPECT_EQ(static_cast<uint>(2), m_it->second.size());
 }
