+++ engine/source/unit_tests/SettingKeybindings_test.cpp
@@ -7,7 +7,7 @@ TEST(SW_Engine_SettingKeybindings, empty_setting)
   SettingKeybindings sk;
   vector<string> bindings = sk.get_keybindings(setting);
 
-  EXPECT_EQ(0, bindings.size());
+  EXPECT_EQ(static_cast<size_t>(0), bindings.size());
 }
 
 TEST(SW_Engine_SettingKeybindings, parse_single)
@@ -17,7 +17,7 @@ TEST(SW_Engine_SettingKeybindings, parse_single)
   SettingKeybindings sk;
   vector<string> bindings = sk.get_keybindings(setting);
 
-  EXPECT_EQ(1, bindings.size());
+  EXPECT_EQ(static_cast<size_t>(1), bindings.size());
   EXPECT_EQ("123", bindings.at(0));
 }
 
@@ -28,7 +28,7 @@ TEST(SW_Engine_SettingKeybindings, parse_multiple)
   SettingKeybindings sk;
   vector<string> bindings = sk.get_keybindings(setting);
 
-  EXPECT_EQ(3, bindings.size());
+  EXPECT_EQ(static_cast<size_t>(3), bindings.size());
   EXPECT_EQ("123", bindings.at(0));
   EXPECT_EQ("343", bindings.at(1));
   EXPECT_EQ("1222", bindings.at(2));
