+++ world/magic/source/unit_tests/SpellKnowledge_test.cpp
@@ -6,11 +6,11 @@ TEST(SW_World_Magic_SpellKnowledge, store_values)
   IndividualSpellKnowledge isk;
   isk.set_castings(15);
 
-  EXPECT_EQ(0, sk.get_spell_knowledge("blast").get_castings());
+  EXPECT_EQ(static_cast<uint>(0), sk.get_spell_knowledge("blast").get_castings());
 
   sk.set_spell_knowledge("big_blast", isk);
 
-  EXPECT_EQ(15, sk.get_spell_knowledge("big_blast").get_castings());
+  EXPECT_EQ(static_cast<uint>(15), sk.get_spell_knowledge("big_blast").get_castings());
 }
 
 TEST(SW_World_Magic_SpellKnowledge, serialization_id)
@@ -63,14 +63,14 @@ TEST(SW_World_Magic_SpellKnowledge, count_spells_known)
   isk.set_castings(3);
   isk.set_bonus(4);
 
-  EXPECT_EQ(0, sk.count_spells_known());
+  EXPECT_EQ(static_cast<uint>(0), sk.count_spells_known());
 
   sk.set_spell_knowledge("frobozz", isk);
 
-  EXPECT_EQ(1, sk.count_spells_known());
+  EXPECT_EQ(static_cast<uint>(1), sk.count_spells_known());
 
   sk.set_spell_knowledge("blaze", isk);
   sk.set_spell_knowledge("pontificate", isk);
 
-  EXPECT_EQ(3, sk.count_spells_known());
+  EXPECT_EQ(static_cast<uint>(3), sk.count_spells_known());
 }
